change layout/add stage name status view/some fixed
git-svn-id: https://svn.osdn.net/svnroot/switch/airSwitch/trunk@489 6544d933-5e26-41f1-b2f0-1bd7b5b7622e

+++ src/MediaSyncronizer.as
@@ -168,9 +168,12 @@
 							name = name.substr(12);
 						}
 						var file:File = getFile(name);
-						if (file.exists) {
-							file.deleteFile();
-							trace("delete: " + file.nativePath);
+						try {
+							if (file && file.exists) {
+								file.deleteFile();
+								trace("delete: " + file.nativePath);
+							}
+						} catch (error:Error) {
 						}
 					}
 				}
